# Use smartstring for name of field (#819).