Replaced name of field with a `smartstring` instead of `String`. Running `cargo bench --features expose-test-schema` indicates a slight improvement on the `query_type_name` test and slightly lower performance on the `introspection_query` test. Here is an example of two runs:

**Master**:
test introspection_query ... bench:     962,848 ns/iter (+/- 70,973)
test query_type_name     ... bench:      11,557 ns/iter (+/- 894)

**This PR/branch**:
test introspection_query ... bench:     978,076 ns/iter (+/- 70,523)
test query_type_name     ... bench:      11,354 ns/iter (+/- 1,471)